test: test 위치 변경 및 새로운 테스트 추가
